requirements.txt
transformers>=4.40.0
torch>=2.2.0
accelerate>=0.29.0
python-dotenv>=1.0.1
pandas>=2.2.0
matplotlib>=3.8.0
openai>=1.30.0
